Update selector with correct schema
