+++ src/Panel.tsx
@@ -3,7 +3,8 @@ import React, { useCallback } from "react";
 
 import { ADDON_ID, PANEL_ID, START_BUILD } from "./constants";
 import { Authentication } from "./screens/Authentication/Authentication";
-import { LinkedProject, LinkProject } from "./screens/LinkProject/LinkProject";
+import { LinkedProject } from "./screens/LinkProject/LinkedProject";
+import { LinkProject } from "./screens/LinkProject/LinkProject";
 import { VisualTests } from "./screens/VisualTests/VisualTests";
 import { AddonState } from "./types";
 import { client, Provider, useAccessToken } from "./utils/graphQLClient";
@@ -53,7 +54,11 @@ export const Panel = ({ active }: PanelProps) => {
     );
 
   if (projectId && projectIdChanged) {
-    return <LinkedProject projectId={projectId} goToNext={clearProjectIdChanged} />;
+    return (
+      <Provider key={PANEL_ID} value={client}>
+        <LinkedProject projectId={projectId} goToNext={clearProjectIdChanged} />
+      </Provider>
+    );
   }
   return (
     <Provider key={PANEL_ID} value={client}>
