+++ src/screens/LinkProject/LinkProject.stories.tsx
@@ -6,7 +6,8 @@ import React from "react";
 import { Build, ProjectQueryQuery, SelectProjectsQueryQuery } from "../../gql/graphql";
 import { storyWrapper } from "../../utils/graphQLClient";
 import { withFigmaDesign } from "../../utils/withFigmaDesign";
-import { LinkedProject, LinkProject } from "./LinkProject";
+import { LinkedProject } from "./LinkedProject";
+import { LinkProject } from "./LinkProject";
 
 const meta = {
   component: LinkProject,
@@ -20,38 +21,40 @@ const meta = {
         graphql.query("SelectProjectsQuery", (req, res, ctx) =>
           res(
             ctx.data({
-              accounts: [
-                {
-                  id: "account:123",
-                  name: "yummly",
-                  projects: [
-                    {
-                      id: "123",
-                      name: "optics",
-                      webUrl: "https://www.chromatic.com/builds?appId=123",
-                      code: "randomcode",
-                    },
-                    {
-                      id: "456",
-                      name: "design-system",
-                      webUrl: "https://www.chromatic.com/builds?appId=456",
-                      code: "randomcode",
-                    },
-                  ],
-                },
-                {
-                  id: "account:456",
-                  name: "acme corp",
-                  projects: [
-                    {
-                      id: "789",
-                      name: "acme",
-                      webUrl: "https://www.chromatic.com/builds?appId=789",
-                      code: "randomcode",
-                    },
-                  ],
-                },
-              ],
+              viewer: {
+                accounts: [
+                  {
+                    id: "account:123",
+                    name: "yummly",
+                    projects: [
+                      {
+                        id: "123",
+                        name: "optics",
+                        webUrl: "https://www.chromatic.com/builds?appId=123",
+                        projectToken: "randomcode",
+                      },
+                      {
+                        id: "456",
+                        name: "design-system",
+                        webUrl: "https://www.chromatic.com/builds?appId=456",
+                        projectToken: "randomcode",
+                      },
+                    ],
+                  },
+                  {
+                    id: "account:456",
+                    name: "acme corp",
+                    projects: [
+                      {
+                        id: "789",
+                        name: "acme",
+                        webUrl: "https://www.chromatic.com/builds?appId=789",
+                        projectToken: "randomcode",
+                      },
+                    ],
+                  },
+                ],
+              },
             } satisfies SelectProjectsQueryQuery)
           )
         ),
@@ -78,20 +81,14 @@ const meta = {
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Default: Story = {
+export const SelectProject: Story = {
   parameters: withFigmaDesign(
     "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=304-318196&t=3EAIRe8423CpOQWY-4"
   ),
 };
 
-export const Adding: Story = {
-  parameters: withFigmaDesign(
-    "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=304-318252&t=3EAIRe8423CpOQWY-4"
-  ),
-};
-
 export const Linked: Story = {
-  render: () => <LinkedProject projectId="789" />,
+  render: () => <LinkedProject projectId="789" goToNext={action("goToNext")} />,
   parameters: {
     ...withFigmaDesign(
       "https://www.figma.com/file/GFEbCgCVDtbZhngULbw2gP/Visual-testing-in-Storybook?type=design&node-id=330-472759&t=3EAIRe8423CpOQWY-4"
